requirements.txt
# --- Web framework ---
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
strawberry-graphql[fastapi]>=0.260.0

# --- Database (MongoDB) ---
beanie>=2.1.0
motor>=3.6.0

# --- Auth ---
pyjwt>=2.9.0
passlib[bcrypt]>=1.7.4

# --- Cache ---
redis>=5.1.1

# --- BigQuery (used by property feature) ---
google-cloud-bigquery>=3.26.0

# --- Config / utilities ---
pydantic[email]>=2.9.0
pydantic-settings>=2.5.2
python-dotenv>=1.0.1
